+++ sdl2/test/sdl2ext_draw_test.py
@@ -109,6 +109,13 @@ def test_line(testsurf):
     assert all(x == 255 for x in view[-1][-1][:3])
     assert all(x == 0 for x in view[1][5][:3])
 
+    # Test surfaces with nonstandard bpp values
+    fmts = ["RGB332", "RGBA4444"]
+    for f in fmts:
+        sf2 = _create_surface((10, 10), fmt=f)
+        sdl2ext.line(sf2.contents, WHITE, (1, 1, 9, 9))
+        SDL_FreeSurface(sf2)
+
     # Test exception on bad input
     with pytest.raises(ValueError):
         sdl2ext.line(sf.contents, WHITE, (1, 2, 3))